ocaml

+++ build-packages.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
-set -e
+if test -f "Packages.bz2"; then
+    rm Packages.bz2
+fi
 dpkg-scanpackages -m ./debs > Packages
-rm Packages.bz2
-bzip2 Packages
+bzip2 -k Packages
